requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
dashscope>=1.20.0
python-dotenv>=1.0.0
pydantic>=2.0
requests>=2.31.0